Location first class (#3)
* Create Location table

And populate it based on the cross join of Latitude and Longitude tables.

Additionally adding a location record of "~~" which holds NULL for both Latitude and Longitude values to signify that this location cannot be specified.



* Apply Location to Island Table



Also changed where the `PRAGMA foreign_keys = ON;` call is made to be at the beginning.e



* Use Location in HeartContainer table



Also update the island-tasks query to utilize the location for Island and HeartContainer.



Doesn't feel right, but also add the "various" Island which is essentially location-less…



* Backpedal on adding the `various` Island



An island should not be able to have a "NULL" location.

A Location should be able to have no latitude or longitude, an Island needs a place.

Therefore, we leave the decision of a "default location name" up to the individual queries that wish to name them.



* Use Location in HeartPiece table



Also update the Island Tasks query



* Use Location in the Item table



Update island-tasks query for Required and Optional items.



* Use Location in TreasureChart table



Update island-tasks query



* Use Location in TriforceChart table



Update island-task query



* Ideal location setup (#4)



* This is the ideal location setup



However HUGE caveat of time to run the query with all of the correlated subqueries



* Use union all for NULL Location features



This dramatically speeds up the query