Add OtherChart Table (#7)
Also a SunkenChart table for mapping a SunkenTreasure to an OtherChart.
A problem I ran into with this was that I wanted to be sure the `location` of the SunkenTreasure and OtherChart were identical before establishing the mapping between the two. Because, the OtherChart and SunkenTreasure location MUST be the same, and there is no easy way to enforce this, I came up with a trigger on the table before insert to check that the locations were the same.

I tried also having a location field on the SunkenChart table itself with foreign keys to both the SunkenTreasure location and the OtherChart location. But this is not allowed since the `location` field in both of those tables are not primary keys nor have unique constraints.